refactoring crud demo
Former-commit-id: d415a3bc8618e2074ff4db6aae129a1a1607f35c

+++ gin-project/dao/model.go
@@ -2,13 +2,13 @@ package dao
 
 // User Model.
 type User struct {
-	ID   uint   `gorm:"primary_key"`
+	ID   uint64 `gorm:"primary_key"`
 	Name string `gorm:"size:255"` // string默认长度为255, 使用这种tag重设。
 }
 
 // Student model for test.
 type Student struct {
-	ID   uint   `gorm:"primary_key"`
+	ID   uint64 `gorm:"primary_key"`
 	Name string `gorm:"size:255"`
-	Age  uint
+	Age  uint64
 }
